Moved validate input arguments to point directory
